requirements.txt
apistar==0.7.2
asgiref==3.2.10
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.0
daphne==2.5.0
Django==3.1
djangorestframework==3.11.1
hyperlink==20.0.1
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2==2.7.6.1
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
typesystem==0.2.4
uritemplate==3.0.1
urllib3==1.25.10
zipp==3.1.0
zope.interface==5.1.0
